Log4j properties now is supported in the generator, i think is very important to have output to know how the generator is reading the model, also i've added the cdb manager xml file to templates for the generation of a simple manager for acs.
I've updated the model file, cause there's some problems, for i.e.: functions positions() or variables positions can be exists due there's exists the position variable, is not so important but is one of the restrictions of the model design also the IDL compilaton throw a problem with that.

Added a validation in the template container, i mean, if the container is java and the instance component is not java, the template will not add that instance.

Validation for the arguments, check if the files exists, if the module choosed exist in the model.

The components generated by default are under the alma package.