IDL templates files are fixed and more things, an <<ETopicError>> has been defined in the EProfile file, this stereotype contains the tags description and shortDescription and his metamodel or base element is uml::Signal.
Automatic comment support is setted in the templates, also the comment function allows to generate or add the comments specified in the model, this comment are wrapped by a java function which is called in the xtend common extension file.

A new util class under the xtend packages has been added, this class is for helper or util static functions to support the code generation, an example of this is the comment wrapper line, due the model developer can comment several words or line, in this class was implemented a comment line wrapper for function comments, variable's comments, etc...etc....

The idl generation is fully supported now, only left the dependency to order the structs in the IDL common file generated.