Error expection can be specified in the model, this errors are generated as a <Module>Err.xml.
Several function to get the date, autor for the make file comments.

Was implemented a topological sort to set the correct position of structs in the <Module>Common.idl file, this algorith is developed for the generator to sort out the order of depencency of : structs, typedef sequences<>..., and hierarchical IDL. This algorithms implemented in java and are executed by *.ext functions which are used in the template, yes the templates can call java functions (only statics functions) to process special data, in this case the order of the structs, etc... In the templates there is nothing to change with this implementation, i mean, everything is setup in the *.ext funcions, see the common.ext file and the function getETopics(...). I think that this implementation is very important meanwhile there is no "native" solution in ACS or TAO compilers, so, this can be a university project -> implement a mechanism who reads the any IDL file and check his order, if anything is not fine (like circular dependency) sort out the order of anything (IDL structs for example) before compile IDL's.

A not so important solution is the comments in the functions and vars in the files generated, this comment can be set in the model, but the generator automatically cut off the line when this reach 10 words, this number can be specified in the BaseStaticConfig class, so you can put huge line of comments and the generator will put them in the right way.