This version of the prototype model produces a definition clash The generated IDL (with verison 135 of the generator) puts all type/classes directly in the
          prototype

namespace.

The original prototype code that I have tried to reproduce was instead using the

          DEV_M, OS_M and SUB_M

namespaces and was putting the definition of some enumerations directly
in the classes.

Therefore there are definitions that are in different namespaces in the original prototype and that fall
together in the generated code, clashing:

            DEV_M::Lamp::m_mode
            DEV_M::Motor:m_mode

We will have to discuss what conventions we should use in this respect and what should be supported
by the generator.