Again the and build script was presenting problems on build a successfuly jar file. This was fixed.
The example folder doesn't exists anymore, now is called test, in which - have the output, xmi, and model folder. Each of these folder contains in separated folder model information name in the same way of the model. The output folder contains the code generated, xmi folder the xmi files exported from MD, and models the models to generate, test, etc. Also the test folder has an Makefile to remove the MD autogenerated backup files, remove the xmi files, and remove the code generate and the testing generation.

Several examples must be updated now, also the makefile to test. This testing depends on the successfuly code generation.

Warning: the code generation is not tested yet.